修改好Completer的Caltech和Mnist的参数

+++ dataset/voc.py
@@ -15,7 +15,7 @@ class Voc(Dataset):
         self.convert_type(standard)
 
     def convert_type(self, standard):
-        self.targets = self.targets.astype('int16')
+        self.targets = self.targets.astype('int64')
         self.view_0 = torch.from_numpy(self.view_0.astype('float32'))
         self.view_1 = torch.from_numpy(self.view_1.astype('float32'))
         self.view_0 = Standard(self.view_0, standard)
